requirements.txt
Flask==2.2.3
keras==2.10.0
pandas==1.4.3
Pillow==9.4.0
tensorboard==2.10.1
tensorflow==2.10.0